Accounting Create Model in Db Almost Done...

+++ Accounting.App/Customer/customersFrm.cs
@@ -7,6 +7,8 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using System.Xml.Linq;
+using Accounting.App.Customer;
 using Accounting.DataLayer.Context;
 
 namespace Accounting.App
@@ -58,19 +60,49 @@ namespace Accounting.App
 
         private void removeCustomerBtn_Click(object sender, EventArgs e)
         {
+            var customer = int.Parse(customersDatagv.CurrentRow.Cells[0].Value.ToString());
             if (customersDatagv.CurrentRow != null)
             {
                 using (UnitOfWork db =new UnitOfWork())
                 {
-                    var customer = int.Parse(customersDatagv.CurrentRow.Cells[0].Value.ToString());
-                    db.CustomersRepository.RemoveCustomer(customer);
-                    db.Save();
-                    GridViewDataBind();
+                    var name = customersDatagv.CurrentRow.Cells[1].Value.ToString();
+                    if (RtlMessageBox.Show(
+                            
+                            $"آیا از حدف{name} مطمئن هستید؟ " ,"توجه",
+                            MessageBoxButtons.YesNo,
+                            MessageBoxIcon.Warning)==DialogResult.Yes)
+                    {
+                       
+                        db.CustomersRepository.RemoveCustomer(customer);
+                        db.Save();
+                        GridViewDataBind();
+                    }
                 }
             }
             else
             {
-                MessageBox.Show("اگر قصد حذف مشتری را دارید باید انتخابش کنید!");
+                RtlMessageBox.Show("اگر قصد حذف مشتری را دارید باید انتخابش کنید!");
+            }
+        }
+
+        private void addCustomerBtn_Click(object sender, EventArgs e)
+        {
+            addOrEditFrm addForm = new addOrEditFrm();
+            if(addForm.ShowDialog() == DialogResult.OK) 
+                GridViewDataBind();
+        }
+
+        private void editCustomerBtn_Click(object sender, EventArgs e)
+        {
+            if (customersDatagv.CurrentRow != null)
+            {
+                var customerId = int.Parse(customersDatagv.CurrentRow.Cells[0].Value.ToString());
+                    addOrEditFrm editForm = new addOrEditFrm();
+                    editForm.CustomerId = customerId;
+                    if (editForm.ShowDialog() == DialogResult.OK)
+                    {
+                        GridViewDataBind();
+                    }
             }
         }
     }
